Rearrange drivers into upper directory
