+++ kernel/src/drivers/mod.rs
@@ -1,4 +1,6 @@
-pub mod virtio_net;
+mod device_tree;
+mod bus;
+mod net;
 
 use lazy_static::lazy_static;
 use alloc::prelude::*;
@@ -10,4 +12,8 @@ pub trait Driver : Send {
 
 lazy_static! {
     pub static ref DRIVERS: SpinNoIrqLock<Vec<Box<Driver>>> = SpinNoIrqLock::new(Vec::new());
+}
+
+pub fn init(dtb: usize) {
+    device_tree::init(dtb);
 }
