+++ kernel/src/drivers/net/mod.rs
@@ -1,0 +1,1 @@
+pub mod virtio_net;
